+++ libraries/SysAPI/include/sys/Pipe.hpp
@@ -21,6 +21,8 @@ public:
   fs::File &write_file() { return m_write_file; }
   const fs::File &write_file() const { return m_write_file; }
 
+  static void make_fifo(var::StringView path, fs::Permissions permissions = fs::Permissions());
+
 private:
   fs::File m_read_file;
   fs::File m_write_file;